+++ packages/audioplayers/example/integration_test/platform_features.dart
@@ -9,6 +9,7 @@ class PlatformFeatures {
     hasPlaylistSourceType: false,
     hasLowLatency: false,
     hasReleaseModeRelease: false,
+    hasForceSpeaker: false,
     hasDuckAudio: false,
     hasRespectSilence: false,
     hasStayAwake: false,
@@ -27,7 +28,6 @@ class PlatformFeatures {
     hasPlaylistSourceType: false,
     hasReleaseModeRelease: false,
     hasLowLatency: false,
-    hasDuckAudio: false,
     hasBalance: false,
   );
 
@@ -36,6 +36,7 @@ class PlatformFeatures {
     hasPlaylistSourceType: false,
     hasLowLatency: false,
     hasReleaseModeRelease: false,
+    hasForceSpeaker: false,
     hasDuckAudio: false,
     hasRespectSilence: false,
     hasStayAwake: false,
@@ -51,6 +52,7 @@ class PlatformFeatures {
     // MP3 duration is estimated: https://bugzilla.gnome.org/show_bug.cgi?id=726144
     // Use GstDiscoverer to get duration before playing: https://gstreamer.freedesktop.org/documentation/pbutils/gstdiscoverer.html?gi-language=c
     hasMp3Duration: false,
+    hasForceSpeaker: false,
     hasDuckAudio: false,
     hasRespectSilence: false,
     hasStayAwake: false,
@@ -63,6 +65,7 @@ class PlatformFeatures {
     hasPlaylistSourceType: false,
     hasLowLatency: false,
     hasReleaseModeRelease: false,
+    hasForceSpeaker: false,
     hasDuckAudio: false,
     hasRespectSilence: false,
     hasStayAwake: false,
@@ -85,6 +88,7 @@ class PlatformFeatures {
   final bool hasMp3Duration;
 
   final bool hasPlaybackRate;
+  final bool hasForceSpeaker;
   final bool hasDuckAudio; // Not yet tested
   final bool hasRespectSilence; // Not yet tested
   final bool hasStayAwake; // Not yet tested
@@ -109,6 +113,7 @@ class PlatformFeatures {
     this.hasBalance = true,
     this.hasSeek = true,
     this.hasPlaybackRate = true,
+    this.hasForceSpeaker = true,
     this.hasDuckAudio = true,
     this.hasRespectSilence = true,
     this.hasStayAwake = true,