+++ packages/audioplayers/example/lib/components/player_widget.dart
@@ -6,10 +6,7 @@ import 'package:flutter/material.dart';
 class PlayerWidget extends StatefulWidget {
   final AudioPlayer player;
 
-  const PlayerWidget({
-    super.key,
-    required this.player,
-  });
+  const PlayerWidget({super.key, required this.player});
 
   @override
   State<StatefulWidget> createState() {
@@ -18,19 +15,21 @@ class PlayerWidget extends StatefulWidget {
 }
 
 class _PlayerWidgetState extends State<PlayerWidget> {
-  PlayerState? _audioPlayerState;
+  PlayerState? _playerState;
   Duration? _duration;
   Duration? _position;
 
-  PlayerState _playerState = PlayerState.stopped;
   StreamSubscription? _durationSubscription;
   StreamSubscription? _positionSubscription;
   StreamSubscription? _playerCompleteSubscription;
   StreamSubscription? _playerStateChangeSubscription;
 
   bool get _isPlaying => _playerState == PlayerState.playing;
+
   bool get _isPaused => _playerState == PlayerState.paused;
+
   String get _durationText => _duration?.toString().split('.').first ?? '';
+
   String get _positionText => _position?.toString().split('.').first ?? '';
 
   AudioPlayer get player => widget.player;
@@ -38,6 +37,18 @@ class _PlayerWidgetState extends State<PlayerWidget> {
   @override
   void initState() {
     super.initState();
+    // Use initial values from player
+    _playerState = player.state;
+    player.getDuration().then(
+          (value) => setState(() {
+            _duration = value;
+          }),
+        );
+    player.getCurrentPosition().then(
+          (value) => setState(() {
+            _position = value;
+          }),
+        );
     _initStreams();
   }
 
@@ -61,6 +72,7 @@ class _PlayerWidgetState extends State<PlayerWidget> {
 
   @override
   Widget build(BuildContext context) {
+    final color = Theme.of(context).primaryColor;
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: <Widget>[
@@ -72,21 +84,21 @@ class _PlayerWidgetState extends State<PlayerWidget> {
               onPressed: _isPlaying ? null : _play,
               iconSize: 48.0,
               icon: const Icon(Icons.play_arrow),
-              color: Colors.cyan,
+              color: color,
             ),
             IconButton(
               key: const Key('pause_button'),
               onPressed: _isPlaying ? _pause : null,
               iconSize: 48.0,
               icon: const Icon(Icons.pause),
-              color: Colors.cyan,
+              color: color,
             ),
             IconButton(
               key: const Key('stop_button'),
               onPressed: _isPlaying || _isPaused ? _stop : null,
               iconSize: 48.0,
               icon: const Icon(Icons.stop),
-              color: Colors.cyan,
+              color: color,
             ),
           ],
         ),
@@ -114,7 +126,7 @@ class _PlayerWidgetState extends State<PlayerWidget> {
                   : '',
           style: const TextStyle(fontSize: 16.0),
         ),
-        Text('State: $_audioPlayerState'),
+        Text('State: ${_playerState ?? '-'}'),
       ],
     );
   }
@@ -138,7 +150,7 @@ class _PlayerWidgetState extends State<PlayerWidget> {
     _playerStateChangeSubscription =
         player.onPlayerStateChanged.listen((state) {
       setState(() {
-        _audioPlayerState = state;
+        _playerState = state;
       });
     });
   }