+++ packages/audioplayers/example/lib/components/tab_content.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class TabContent extends StatelessWidget {
+  final List<Widget> children;
+
+  const TabContent({super.key, required this.children});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Container(
+        alignment: Alignment.topCenter,
+        child: SingleChildScrollView(
+          controller: ScrollController(),
+          child: Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8),
+            child: Column(
+              children: children,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}