+++ packages/audioplayers_android/android/src/main/kotlin/xyz/luan/audioplayers/player/MediaPlayerPlayer.kt
@@ -12,13 +12,15 @@ class MediaPlayerPlayer(
     private val mediaPlayer = createMediaPlayer(wrappedPlayer)
 
     private fun createMediaPlayer(wrappedPlayer: WrappedPlayer): MediaPlayer {
-        return MediaPlayer().apply {
+        val mediaPlayer = MediaPlayer().apply {
             setOnPreparedListener { wrappedPlayer.onPrepared() }
             setOnCompletionListener { wrappedPlayer.onCompletion() }
             setOnSeekCompleteListener { wrappedPlayer.onSeekComplete() }
             setOnErrorListener { _, what, extra -> wrappedPlayer.onError(what, extra) }
             setOnBufferingUpdateListener { _, percent -> wrappedPlayer.onBuffering(percent) }
         }
+        wrappedPlayer.context.setAttributesOnPlayer(mediaPlayer)
+        return mediaPlayer
     }
 
     override fun getDuration(): Int? {
@@ -41,7 +43,7 @@ class MediaPlayerPlayer(
     override fun setRate(rate: Float) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
             mediaPlayer.playbackParams = mediaPlayer.playbackParams.setSpeed(rate)
-        } else {
+        } else if (rate != 1.0f) {
             error("Changing the playback rate is only available for Android M/23+ or using LOW_LATENCY mode.")
         }
     }
@@ -77,8 +79,6 @@ class MediaPlayerPlayer(
     }
 
     override fun updateContext(context: AudioContextAndroid) {
-        // TODO(luan) is this global?
-        wrappedPlayer.audioManager.isSpeakerphoneOn = context.isSpeakerphoneOn
         context.setAttributesOnPlayer(mediaPlayer)
         if (context.stayAwake) {
             mediaPlayer.setWakeMode(wrappedPlayer.applicationContext, PowerManager.PARTIAL_WAKE_LOCK)