+++ packages/audioplayers_android/android/src/main/kotlin/xyz/luan/audioplayers/player/WrappedPlayer.kt
@@ -18,6 +18,7 @@ class WrappedPlayer internal constructor(
     private val ref: AudioplayersPlugin,
     val playerId: String,
     var context: AudioContextAndroid,
+    private val soundPoolManager: SoundPoolManager,
 ) {
     private var player: Player? = null
 
@@ -122,7 +123,21 @@ class WrappedPlayer internal constructor(
             focusManager.handleStop()
         }
         this.context = audioContext.copy()
-        player?.updateContext(context)
+
+        // AudioManager values are set globally
+        audioManager.mode = context.audioMode
+        audioManager.isSpeakerphoneOn = context.isSpeakerphoneOn
+
+        player?.let { p ->
+            p.stop()
+            prepared = false
+            // Context is only applied, once the player.reset() was called
+            p.updateContext(context)
+            source?.let {
+                p.setSource(it)
+                p.configAndPrepare()
+            }
+        }
     }
 
     // Getters
@@ -149,7 +164,7 @@ class WrappedPlayer internal constructor(
         get() = ref.getApplicationContext()
 
     val audioManager: AudioManager
-        get() = applicationContext.getSystemService(Context.AUDIO_SERVICE) as AudioManager
+        get() = ref.getAudioManager()
 
     /**
      * Playback handling methods
@@ -178,7 +193,7 @@ class WrappedPlayer internal constructor(
         }
         if (releaseMode != ReleaseMode.RELEASE) {
             pause()
-            if(prepared) {
+            if (prepared) {
                 if (player?.isLiveStream() == true) {
                     player?.stop()
                     prepared = false
@@ -268,8 +283,9 @@ class WrappedPlayer internal constructor(
         return false
     }
 
+    @Suppress("UNUSED_PARAMETER")
     fun onBuffering(percent: Int) {
-        // TODO(luan) expose this as a stream
+        // TODO(luan): expose this as a stream
     }
 
     fun onSeekComplete() {
@@ -286,7 +302,7 @@ class WrappedPlayer internal constructor(
     private fun createPlayer(): Player {
         return when (playerMode) {
             MEDIA_PLAYER -> MediaPlayerPlayer(this)
-            LOW_LATENCY -> SoundPoolPlayer(this)
+            LOW_LATENCY -> SoundPoolPlayer(this, soundPoolManager)
         }
     }